third_party/blink/web_tests/platform/generic/fast/lists/ol-reversed-dynamic-simple-expected.txt

This tests that changing the "reversed" property updates markers.

5 Five
4 Four
3 Three
2 Two
1 One


This tests that changing the "reversed" attribute updates markers.

5 Five
4 Four
3 Three
2 Two
1 One


This tests that adding new items to a reversed list updates markers.

7 Seven
6 Six
5 Five
-> 4 Four
3 Three
2 Two
1 One


This tests that removing existing tems from a reversed list updates markers.

5 Five
4 Four
3 Three
2 Two
1 One
